requirements.txt

celery==5.4.0
Django==5.0.7
django-admin==2.0.2
django-environ==0.11.2
redis==5.0.8
requests==2.32.3

